nixpkgs: implement new postgresqlPackages infrastructure
This adds five new top-level attrsets:

    pkgs.postgresql93Packages
    pkgs.postgresql94Packages
    pkgs.postgresql95Packages
    pkgs.postgresql96Packages
    pkgs.postgresql10Packages

Each of these contains a set of PostgreSQL extensions, built against the
particular specified postgresql package. This makes it far easier for specific
PostgreSQL extensions to be built and used in NixOS modules, and provides an
easier way to test extension builds against multiple versions. For example, we
can now use either:

    pkgs.postgresql96Packages.timescaledb

or

    pkgs.postgresql10Packages.timescaledb

when building and specifying extensions.

By default, the prior top-level attribute names such as 'timescaledb' HAVE NOT
been removed. They will be deprecated in a future commit to encourage people to
move to explicitly-versioned attrsets.

Signed-off-by: Austin Seipp <aseipp@pobox.com>