nixpkgs: break weird cyclic postgresql dependency via gdal
See the comment in gdal's expression. Essentially, postgis depends on gdal, and
gdal depends on postgresql. This means if you use any other version of
PostgreSQL (e.g. 9.3) other than the default (e.g. 10), the former is dependent
on the latter, because 'gdal' wants libpq.

The reason for this feature is so that GDAL can use libpq to query for vector
data, primarily of PostGIS types. Thus, when building postgis itself, we build
a private copy of gdal that excludes this dependency.

Signed-off-by: Austin Seipp <aseipp@pobox.com>