postgresqlPackages: add versionCheck passthru to extension builds for checking PostgreSQL compatibility
Not every PostgreSQL extension builds with every version we have available; for
instance, timescaledb is only available for PostgreSQL 9.6 and later.

To do this, extensions may use a 'passthru.versionCheck' attribute, which
should return 'true' if the extension can be built and used against the given
postgresql version. This avoids needless build failures and helps indicate
to the user that an extension is not available.

Unfortunately the PostgreSQL expression did not previously offer a 'version'
attribute to check against, so we must add it (as expected). This implies a
rebuild of the PostgreSQL derivations, unfortunately.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
(cherry picked from commit 4eaf18bdfb3bfb9a1b8c551a1a33b8dee0adfea9)