postgresql11: keep bitcode/llvmjit.so in $out, not $lib
With JIT support enabled, the Postgres build system keeps a copy of
bitcode for the server, and a new llvmjit.so extension, in the $lib
output. However, this bloats the closure of any libpq-dependent
libraries by requiring a hard dependency on libLLVM.so, as well as the
actual bitcode files (which aren't useful at all to clients).

Moving both of these into $out does not hurt binary output size (since
it already was dependent on $lib anyway), but does help all clients.

Signed-off-by: Austin Seipp <aseipp@pobox.com>