postgresqlPackages.postgis: fix gdal against the same postgresql the extension is built against
GDAL's Postgres support had a weird cyclic dependency: GDAL depends on
Postgres (for possible spatial information querying for GDAL clients),
and PostGIS depends on GDAL. This sounds fine, except it has a dirty
follow up: GDAL is *only* built against the default version of
PostgreSQL. This means if you try to use a non-default version of
Postgres with PostGIS, it pulls in GDAL, which pulls in some other
version of Postgres. This results in weird situations like PostgreSQL
9.4 + PostGIS depending on PostgreSQL 9.6 in their closure.

In a previous experiment I tried making this optional, but there's a
better way: instead of making it optional, we simply let the default
'gdal' expression depend on 'postgresql' (whatever version the default
might be), and then we simply override the version of GDAL used by
postgis to be the *same* postgres that postgis is built against. This
effectively results in a private copy of gdal that only is used by the
postgis extension builds.

Signed-off-by: Austin Seipp <aseipp@pobox.com>