postgresqlPackages.postgis: fix gdal against the same postgresql the extension is built against
This is an alternative to a previous commit which made Postgres support
for GDAL optional; instead of making it optional, we simply let the
default 'gdal' expression depend on 'postgresql' (whatever version the
default might be), and then we simply override the version of GDAL used
by postgis to be the *same* postgres that postgis is built against. This
effectively results in a private copy of gdal that only is used by
the postgis extension builds.

This also effectively reverts the previous uses GDAL changes and makes
postgres non-optional, again.

Signed-off-by: Austin Seipp <aseipp@pobox.com>