WIP: python3Packages.tensorflow: Fix mklSupport
Just like cudaSupport, mklSupport affects the dependencies that need to
be fetched. Using bazelFlags rather than bazelBuildFlags to specify
--config=mkl ensures these dependencies are fetched. This increases the
number of permutations for deps hash. Luckily, the deps don't take as
long to build and hash as tensorflow itself. I've also added a parameter
that allows a user to override the hash more easily in case they want to
use a feature set we're not aware of.